app.py
import os
import json
import joblib
import pandas as pd
import streamlit as st
from typing import Dict, Any
from dotenv import load_dotenv
import os
os.environ["GEMINI_API_KEY"] = "AIzaSyBrxLOojBfr7Z4Iu3ALkWiISqBQY_Ye_Vc"

# =========================================
# FUNGSI CHAT AI (OpenAI & Gemini)
# =========================================
def chat_reply(system_prompt: str, messages: list, api_key: str, model="gpt-4o-mini", provider="openai"):
    """
    Fungsi untuk mengirim percakapan ke OpenAI atau Gemini
    """
    if provider == "openai":
        import openai
        openai.api_key = api_key
        conversation = [{"role": "system", "content": system_prompt}] + messages
        resp = openai.ChatCompletion.create(
            model=model,
            messages=conversation,
            max_tokens=600,
            temperature=0.2,
        )
        return resp.choices[0].message["content"].strip()

    elif provider == "gemini":
        import google.generativeai as genai
        genai.configure(api_key=api_key)
        prompt_text = system_prompt + "\n\n"
        for m in messages:
            role = m["role"]
            content = m["content"]
            prompt_text += f"{role.upper()}: {content}\n"
        model_gemini = genai.GenerativeModel("models/gemini-2.5-flash")
        response = model_gemini.generate_content(prompt_text)
        return response.text.strip()

    else:
        raise ValueError("Provider tidak dikenali. Gunakan 'openai' atau 'gemini'.")

# =========================================
# FUNGSI MODEL
# =========================================
@st.cache_resource
def load_model(path: str):
    return joblib.load(path)

# =========================================
# KONFIGURASI STREAMLIT
# =========================================
st.set_page_config(page_title="Prediksi Harga Mobil + Chat AI", layout="wide")
st.title("🚗 Prediksi Harga Mobil & Chat AI")

# Sidebar
st.sidebar.header("⚙ Konfigurasi")
model_path = st.sidebar.text_input("Path ke model (.pkl)", value="model.pkl")
example_json_path = st.sidebar.text_input("Path contoh fitur (example_features.json)", value="example_features.json")

openai_key_env = st.sidebar.text_input("Masukkan API Key (opsional)", value="", type="password")
ai_provider = st.sidebar.selectbox("Pilih provider AI", ["openai", "gemini"], index=0)

# ✅ Perbaikan bagian ENV KEY
if ai_provider == "gemini":
    use_env_key = st.sidebar.checkbox("Gunakan dari ENV var GEMINI_API_KEY jika kosong", value=True)
else:
    use_env_key = st.sidebar.checkbox("Gunakan dari ENV var OPENAI_API_KEY jika kosong", value=True)

if use_env_key and not openai_key_env:
    if ai_provider == "gemini":
        openai_key_env = os.environ.get("GEMINI_API_KEY", "")
    else:
        openai_key_env = os.environ.get("OPENAI_API_KEY", "")

# Debug opsional: tampilkan apakah key terbaca
# st.sidebar.write("DEBUG KEY:", bool(openai_key_env))

# =========================================
# LOAD MODEL & SCHEMA
# =========================================
example_schema = {}
try:
    with open(example_json_path, "r") as f:
        example_schema = json.load(f)
except Exception as e:
    st.sidebar.warning(f"Gagal load example_features.json: {e}. Gunakan default.")
    example_schema = {
        "model": "Aygo",
        "year": 2017,
        "transmission": "Manual",
        "mileage": 11730,
        "fuelType": "Petrol",
        "tax": 0,
        "mpg": 68.9,
        "engineSize": 1.0
    }

try:
    model = joblib.load("model.pkl")
except Exception as e:
    st.error(f"Gagal memuat model: {e}")
    st.info("Model kemungkinan dibuat dengan versi scikit-learn berbeda atau tidak lengkap.")
    model = None

# =========================================
# LAYOUT UTAMA
# =========================================
left, right = st.columns([1.2, 1.0])

# =========================================
# KIRI: INPUT FITUR & PREDIKSI
# =========================================
with left:
    st.subheader("📊 Input Fitur Mobil")

    with st.form("input_form"):
        col1, col2 = st.columns(2)
        with col1:
            model_name = st.text_input("Model", value=example_schema.get("model", "Aygo"))
            year = st.number_input("Tahun", min_value=1990, max_value=2025, value=int(example_schema.get("year", 2017)))
            transmission = st.selectbox("Transmisi", ["Manual", "Automatic", "Semi-Auto"], index=0)
            fuel_type = st.selectbox("Jenis Bahan Bakar", ["Petrol", "Diesel", "Hybrid", "Electric"], index=0)
        with col2:
            mileage = st.number_input("Jarak Tempuh (Mileage)", min_value=0, value=int(example_schema.get("mileage", 10000)))
            tax = st.number_input("Pajak (£)", min_value=0, value=int(example_schema.get("tax", 0)))
            mpg = st.number_input("MPG (Miles per Gallon)", min_value=0.0, value=float(example_schema.get("mpg", 60.0)))
            engine_size = st.number_input("Ukuran Mesin (Engine Size)", min_value=0.0, value=float(example_schema.get("engineSize", 1.0)))

        submitted = st.form_submit_button("🚗 Prediksi Harga Mobil")

    if submitted:
        input_dict = {
            "model": model_name,
            "year": year,
            "transmission": transmission,
            "mileage": mileage,
            "fuelType": fuel_type,
            "tax": tax,
            "mpg": mpg,
            "engineSize": engine_size
        }

        X_df = pd.DataFrame([input_dict])

        try:
            pred = model.predict(X_df)
            price = float(pred[0])
            st.success(f"💰 Prediksi harga mobil: *£{price:,.2f}*")
            st.session_state["last_prediction"] = {"price": price, "input": input_dict}
        except Exception as e:
            st.error(f"Gagal melakukan prediksi: {e}")

        # Tombol Analisis Otomatis
        if openai_key_env and st.button("🤖 Analisis Otomatis dengan AI"):
            system_prompt = (
                f"Kamu adalah asisten AI yang menjelaskan hasil prediksi harga mobil.\n"
                f"Model memprediksi harga £{price:,.2f} untuk mobil dengan fitur: {json.dumps(input_dict, ensure_ascii=False)}."
            )
            try:
                ai_reply = chat_reply(system_prompt, [], openai_key_env, provider=ai_provider)
                st.info(ai_reply)
            except Exception as e:
                st.error(f"Gagal menghubungi AI: {e}")
        elif not openai_key_env:
            st.warning("Masukkan API Key di sidebar atau file .env untuk menggunakan fitur AI otomatis.")

    if st.checkbox("📋 Tampilkan data input"):
        st.dataframe(pd.DataFrame([{
            "model": model_name,
            "year": year,
            "transmission": transmission,
            "mileage": mileage,
            "fuelType": fuel_type,
            "tax": tax,
            "mpg": mpg,
            "engineSize": engine_size
        }]))

# =========================================
# KANAN: CHAT AI
# =========================================
with right:
    st.subheader("Chat dengan AI")
    if "chat_history" not in st.session_state:
        st.session_state.chat_history = []

    for role, text in st.session_state.chat_history:
        if role == "user":
            st.markdown(f"*Kamu:* {text}")
        else:
            st.markdown(f"*AI:* {text}")

    user_input = st.text_area("Ketik pertanyaan di sini", height=120)
    col1, col2 = st.columns([1, 1])
    with col1:
        ask_btn = st.button("Kirim ke AI")
    with col2:
        clear_btn = st.button("Hapus chat")

    if clear_btn:
        st.session_state.chat_history = []
        st.rerun()

    if ask_btn and user_input.strip():
        st.session_state.chat_history.append(("user", user_input))
        system_prompt = (
            "Kamu adalah asisten AI yang membantu menjelaskan hasil prediksi harga mobil. "
            "Jawabanmu harus singkat, jelas, dan sesuai konteks prediksi."
        )

        last_pred = st.session_state.get("last_prediction", None)
        if last_pred:
            system_prompt += (
                f"\nTerakhir model memprediksi harga £{last_pred['price']:,.2f} "
                f"untuk mobil dengan fitur {json.dumps(last_pred['input'], ensure_ascii=False)}."
            )

        messages = [{"role": "user" if r == "user" else "assistant", "content": t} for r, t in st.session_state.chat_history[-10:]]

        if not openai_key_env:
            st.error("❌ API key belum diatur.")
        else:
            try:
                ai_reply = chat_reply(system_prompt, messages, openai_key_env, provider=ai_provider)
                st.session_state.chat_history.append(("assistant", ai_reply))
                st.rerun()
            except Exception as e:
                st.error(f"Gagal menghubungi AI: {e}")

# =========================================
# FOOTER
# =========================================
st.markdown("---")
st.markdown(
    """
    *Panduan:*
    - Simpan model terlatih sebagai model.pkl (gunakan joblib.dump(model, "model.pkl")).
    - Simpan contoh fitur sebagai example_features.json.
    - Jalankan aplikasi dengan streamlit run app.py.
    - Pilih provider AI di sidebar (OpenAI atau Gemini) dan masukkan API Key.
    - Kamu juga bisa menyimpan key di file .env:
      
      GEMINI_API_KEY=your_key
      OPENAI_API_KEY=your_key
      
    """
)