requirements.txt
streamlit==1.39.0
pandas==2.2.3
scikit-learn==1.5.2
joblib==1.4.2
google-generativeai==0.8.2
numpy==1.26.4
python-dotenv==1.0.1